Adds stubbed Sheets class
In order to do real work, we are likely going to need more than
a single Python file to hold all the functionality.

This creates a stubbed Sheets class, which would handle the
process of dealing with the various Google Sheets we need to
manipulate - both the dashboard and the puzzling sheets.

This class doesn't do anything yet.

A side effect of this change is to introduce the requirements.txt
file for managing bot dependencies.